Added page functionality...
added previous page tab and aligned it to the right with next tab using div. then I used JS function to hide previous tab when on first page as well as hiding the next tab when the end is reached.

Finally I attempted to connect next tab with next page in api to display additional coins. that function is not working yet and needs work.